+++ tests/pinned_drop.rs
@@ -70,10 +70,10 @@ fn self_ty() {
 
             // pat
             match *self {
-                Self { f: _ } => {}
+                Self { f: () } => {}
             }
-            if let Self { f: _ } = *self {}
-            let Self { f: _ } = *self;
+            if let Self { f: () } = *self {}
+            let Self { f: () } = *self;
         }
     }
 
@@ -113,11 +113,11 @@ fn self_ty() {
 
             // pat
             match *self {
-                Self::Struct { f: _ } => {}
+                Self::Struct { f: () } => {}
                 Self::Tuple(_) => {}
                 Self::Unit => {}
             }
-            if let Self::Struct { f: _ } = *self {}
+            if let Self::Struct { f: () } = *self {}
             if let Self::Tuple(_) = *self {}
             if let Self::Unit = *self {}
         }
@@ -138,7 +138,7 @@ fn self_inside_macro_containing_fn() {
     #[pinned_drop]
     impl PinnedDrop for S {
         fn drop(self: Pin<&mut Self>) {
-            let _ = mac!({
+            mac!({
                 impl S {
                     pub fn _f(self) -> Self {
                         self
@@ -213,7 +213,7 @@ fn self_ty_inside_macro_call() {
     }
 
     impl<T: Send> Trait for Struct<T> {
-        type Assoc2 = ();
+        type Assoc2 = u8;
         const ASSOC2: usize = 2;
         fn assoc2() {}
     }